Rename directory myapplication -> sample
