+++ app/src/androidTest/java/com/daydreamapplications/sample/ExampleInstrumentedTest.kt
@@ -1,4 +1,4 @@
-package com.daydreamapplications.myapplication
+package com.daydreamapplications.sample
 
 import androidx.test.platform.app.InstrumentationRegistry
 import androidx.test.ext.junit.runners.AndroidJUnit4
